add the popup function to separate file
