+++ src/index.js
@@ -3,7 +3,9 @@ import Shows from './shows.js';
 import { showsContainer, searchForm, searchFormInput } from './dynamic.js';
 import likes from './images/Empty_Like.svg';
 
-const bookList = document.getElementById('book-list');
+export let showCard; export let
+  commentButton; export let search; export let
+  show;
 const setNewShow = (programList) => {
   let tvShows = [];
   programList.forEach((program) => {
@@ -34,7 +36,8 @@ const getTVShows = async (search) => {
   return newProgramList;
 };
 
-const displayShows = async (search) => {
+// eslint-disable-next-line import/prefer-default-export
+export const displayShows = async (search) => {
   showsContainer.innerHTML = '';
   searchFormInput.placeholder = 'Type Search your favorite TV show or Genre';
   searchFormInput.classList.remove('error_input_show');
@@ -72,55 +75,8 @@ const displayShows = async (search) => {
     commentButton.innerText = 'Comments';
     commentButton.classList.add('comment_button');
     showCard.appendChild(commentButton);
-    // Pop up function
-    const popUp = async (pick) => {
-      // Make the API call to fetch comments for the selected item
-      const commentsResponse = await fetch(`https://api.tvmaze.com/search/shows?q=${search}`);
-      const commentsData = await commentsResponse.json();
-
-      // Store the comments in the pick object
-      pick.comments = commentsData;
-
-      const popUpContainer = document.createElement('article');
-      popUpContainer.id = 'pop-up';
-      popUpContainer.innerHTML = `
-    <div class="pop-container">
-      <div class="pop-up-header">
-        <h2 class="movie-title">${pick.title}</h2>
-        <button id="close-modal-btn">x</button>
-      </div>
-      <div class="pop-up-body">
-        <div class="image-container">
-          <img src='${pick.image.medium}' class="movie-image">
-        </div>
-        <div class="pop-up-content">
-          <div class="left-content">
-            <label>Genre:</label>
-            <div class="movie-data">${pick.genres}</div>
-            <label>RunTime:</label>
-            <div class="movie-data">${pick.country}</div>
-          </div>
-          <div class="right-content">
-            <label>Show Type:</label>
-            <div class="movie-data">${pick.type}</div>
-            <label>Language:</label>
-            <div class="movie-data">${pick.language}</div>
-          </div>
-        </div>
-      </div>
-    </div>
-  `;
-      document.body.append(popUpContainer);
-      const closeBtn = document.getElementById('close-modal-btn');
-      closeBtn.addEventListener('click', () => {
-        popUpContainer.remove();
-      });
-    };
 
     showsContainer.appendChild(showCard);
-    commentButton.addEventListener('click', () => {
-      popUp(show);
-    });
   });
 };
